Move isNonNullObject, isUint8Array into @iov/encoding
