+++ packages/iov-bcp/types/transactions.d.ts
@@ -1,13 +1,5 @@
 import { As } from "type-tagger";
 import { Hash, Preimage } from "./atomicswaptypes";
-/**
- * Checks if data is a non-null object (i.e. matches the TypeScript object type)
- *
- * Only used internally to this package.
- */
-export declare function isNonNullObject(data: unknown): data is object;
-/** Checks if data is an Uint8Array. Note: Buffer is treated as not a Uint8Array */
-export declare function isUint8Array(data: unknown): data is Uint8Array;
 export declare enum Algorithm {
     Ed25519 = "ed25519",
     Secp256k1 = "secp256k1"
